requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
et_xmlfile==2.0.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
openpyxl==3.1.5
playwright==1.58.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.1
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
